+++ apps/admin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.2.2](https://github.com/F3-Nation/f3-nation/compare/admin@1.2.1...admin@1.2.2) (2026-06-11)
+
+
+### Bug Fixes
+
+* **me:** verify JWT signature at handler layer and fix refresh-rotation race ([#400](https://github.com/F3-Nation/f3-nation/issues/400)) ([853eed5](https://github.com/F3-Nation/f3-nation/commit/853eed58d3d1596a3f03b613b517436af871822f))
+
+## [1.2.1](https://github.com/F3-Nation/f3-nation/compare/admin@1.2.0...admin@1.2.1) (2026-06-03)
+
+
+### Bug Fixes
+
+* **admin:** admin portal minor issues ([#386](https://github.com/F3-Nation/f3-nation/issues/386)) ([ef28ce5](https://github.com/F3-Nation/f3-nation/commit/ef28ce50441138c8ba443a82d9b1a00a84e51005))
+
 ## [1.2.0](https://github.com/F3-Nation/f3-nation/compare/admin@1.1.1...admin@1.2.0) (2026-05-31)